// src/components/Navbar/interface.ts
export enum LinkPage {
  popularСars = "/",
  carShowrooms = "/carShowrooms",
  testDrive = "/testDrive",
  sellingSoon = "/sellingSoon",
}

export enum NamePage {
  popularСars = "Популярні авто",
  carShowrooms = "Автосалони",
  testDrive = "Тест-драйви",
  sellingSoon = "Скоро у продажу",
}